telephony: Do not set default subscriptions on shutdown
On some devices the SIM slots are not deactivated at the same time
when the phone is shut down or rebooted. This means that at some point
only one SIM is activated, and all default subscriptions are set to
this SIM (a feature introduced in http://review.cyanogenmod.org/#/c/121239/).

Change-Id: Ief36004619b75530a21b23efc767e1beb1fb0750
Signed-off-by: Louis Popi <theh2o64@gmail.com>
Signed-off-by: theh2o64@gmail.com <theh2o64@gmail.com>